refactor distance comparison plotting
