+++ m3c2/visualization/__init__.py
@@ -1,5 +1,6 @@
 """Visualization utilities and plotting services."""
 
 from .plot_service import PlotService
+from .plot_comparedistances_service import PlotServiceCompareDistances
 
-__all__ = ["PlotService"]
+__all__ = ["PlotService", "PlotServiceCompareDistances"]
